WIP - Consolidating 3 span classes into one.
The span code has been split across 3 classes since the beginning under the assumption that we could avoid having the critical path run through objective-c. In reality, this hasn't panned out, and also there has been drift between all of the classes as to what the exact meaning of some of the data is.

This was a first attempt to consolidate everything, and the unit tests pass, but there are still a number of bugbears, so I'm parking it to look into at a later date.